requirements.txt
fastapi
uvicorn
jinja2
python-multipart
httpx
pydantic
python-dotenv

# SAP BTP Connectivity
cfenv
sap-xssec
requests
